+++ tests/testthat/test-positividad.R
@@ -3,6 +3,88 @@ test_that("Positividad", {
   #Leemos los datos
   datos_covid <- covidmx::datosabiertos
 
+  #Checamos positividad de pcr-----
+  positividad_agrupados <- datos_covid %>%
+    positividad(entidades =  "BAJA CALIFORNIA", quiet = TRUE, tipo_prueba = "PCR",
+                fill_NA = FALSE, remove_inconclusive = FALSE)
+
+  numero_pruebas <- datos_covid %>%
+    numero_pruebas(entidades =  "BAJA CALIFORNIA", tipo_prueba = "PCR", fill_zeros = TRUE)
+
+  numero_pruebas_positivas <- datos_covid %>%
+    numero_pruebas(entidades =  "BAJA CALIFORNIA", tipo_prueba = "PCR", fill_zeros = TRUE,
+                   .grouping_vars = "RESULTADO_LAB")
+  numero_pruebas_positivas <- numero_pruebas_positivas$numero_pruebas %>%
+    dplyr::filter(RESULTADO_LAB == 1)
+
+  #Checamos # pruebas sea
+  ntests <- numero_pruebas$numero_pruebas$n
+  npos   <- numero_pruebas_positivas$n
+
+  expect_equal(positividad_agrupados$positividad$n_pruebas, ntests)
+
+  #Checamos # pruebas positivas sea
+  expect_equal(positividad_agrupados$positividad$n_positivos, npos)
+
+  #Checamos positividad
+  expect_equal(positividad_agrupados$positividad$Positividad,npos/ntests)
+
+  #Checamos positividad de igg-----
+  positividad_agrupados <- datos_covid %>%
+    positividad(entidades =  "BAJA CALIFORNIA", quiet = TRUE, tipo_prueba = "ANTIGENO",
+                fill_NA = FALSE, remove_inconclusive = FALSE)
+
+  numero_pruebas <- datos_covid %>%
+    numero_pruebas(entidades =  "BAJA CALIFORNIA", tipo_prueba = "ANTIGENO", fill_zeros = TRUE)
+
+  numero_pruebas_positivas <- datos_covid %>%
+    numero_pruebas(entidades =  "BAJA CALIFORNIA", tipo_prueba = "ANTIGENO", fill_zeros = TRUE,
+                   .grouping_vars = "RESULTADO_ANTIGENO")
+  numero_pruebas_positivas <- numero_pruebas_positivas$numero_pruebas %>%
+    dplyr::filter(RESULTADO_ANTIGENO == 1)
+
+  #Checamos # pruebas sea
+  ntests <- numero_pruebas$numero_pruebas$n
+  npos   <- numero_pruebas_positivas$n
+
+  expect_equal(positividad_agrupados$positividad$n_pruebas, ntests)
+
+  #Checamos # pruebas positivas sea
+  expect_equal(positividad_agrupados$positividad$n_positivos, npos)
+
+  #Checamos positividad
+  expect_equal(positividad_agrupados$positividad$Positividad,npos/ntests)
+
+  #Checamos positividad de pcr e igg conjunta-----
+  positividad_agrupados <- datos_covid %>%
+    positividad(entidades =  "BAJA CALIFORNIA", quiet = TRUE,
+                fill_NA = FALSE, remove_inconclusive = FALSE,
+                group_by_tipo_prueba = FALSE)
+
+  numero_pruebas <- datos_covid %>%
+    numero_pruebas(entidades =  "BAJA CALIFORNIA", group_by_tipo_prueba = FALSE,
+                   fill_zeros = FALSE)
+
+  numero_pruebas_positivas <- datos_covid %>%
+    numero_pruebas(entidades =  "BAJA CALIFORNIA",  fill_zeros = FALSE,
+                   group_by_tipo_prueba = FALSE,
+                   .grouping_vars = c("RESULTADO_ANTIGENO", "RESULTADO_LAB"))
+  numero_pruebas_positivas <- numero_pruebas_positivas$numero_pruebas %>%
+    dplyr::filter(RESULTADO_ANTIGENO == 1 | RESULTADO_LAB == 1) %>%
+    dplyr::group_by(!!as.symbol("FECHA_SINTOMAS")) %>%
+    dplyr::summarise(n = sum(n))
+
+  #Checamos # pruebas sea
+  ntests <- numero_pruebas$numero_pruebas$n
+  npos   <- numero_pruebas_positivas$n
+
+  expect_equal(positividad_agrupados$positividad$n_pruebas, ntests)
+
+  #Checamos # pruebas positivas sea
+  expect_equal(positividad_agrupados$positividad$n_positivos, npos)
+
+  #Checamos positividad
+  expect_equal(positividad_agrupados$positividad$Positividad, npos/ntests)
 
   #Chequeo de que sólo lea el Baja California correcto-----
   positividad_agrupados <- datos_covid %>%
@@ -45,5 +127,8 @@ test_that("Positividad", {
   #Chequeo del list_name----
   positividad_prueba <- datos_covid %>% positividad(list_name = "Prueba", quiet = TRUE)
   expect_true("Prueba" %in% names(positividad_prueba))
+
+  #Chequeo de mensaje
+  expect_message(positividad(datos_covid))
 
 })